Change from Integer to Long type for id attribute

+++ springBootApp/src/main/java/org/example/springBootApp/springBootExercises/model/User.java
@@ -2,21 +2,21 @@ package org.example.springBootApp.springBootExercises.model;
 
 public class User {
 
-    private Integer id;
+    private Long id;
     private String name;
     private String email;
 
-    public User(Integer id, String name, String email) {
+    public User(Long id, String name, String email) {
         this.id = id;
         this.name = name;
         this.email = email;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
